requirements.txt
streamlit>=0.80.0
pandas>=1.2.3
numpy>=1.20.1
feedgen>=0.9.0
pyyaml>=5.4.1
